datadog: Use "latest" as DEFAULT_AGENT_TAG (#15)
As you can see on the [Google Cloud repository][0], there is no image tagged as "7". The freshest image is "latest", so I changed the default to that.

[0]: https://console.cloud.google.com/artifacts/docker/datadoghq/us/gcr.io/cluster-agent

datadog: Use "latest" as DEFAULT_AGENT_TAG

As you can see on the [Google Cloud repository][0], there is no image
tagged as "7". The freshest image is "latest", so I changed the default
to that.

[0]: https://console.cloud.google.com/artifacts/docker/datadoghq/us/gcr.io/cluster-agent